==> manuscript_analyses/02_percentage_analysis.Rmd <==
---
title: "Abundance Testing and Percentage per sample"
author: "Martin Bontrager"
date: "8 June 2016"
output: html_document
---

```{r global_options, include=FALSE}
knitr::opts_chunk$set(echo=FALSE, warning=FALSE, message=FALSE)
```

In this document I explore the percent composition of different microbiome samples using data that was generated by phyloseq.

```{r input, echo=FALSE}
library(dplyr)
library(reshape2)
set.seed(9373399)

water <- c("BBW", "BRW", "CBW", "EUW", "FRW", "IJW", "LOW", "MAW", "MMW", "POW", "RMW", "SCW", "TXW", "VIW")
copepod <- c("BRE", "BBE", "CBE", "EUE", "FRE", "IJE", "LOE", "MAE", "MME", "POE", "RME", "SCE", "TXE", "VIE")
phylumTax <- read.csv("output/sample_matrix_phylum.csv")
classTax <- read.csv("output/sample_matrix_class.csv")
orderTax <- read.csv("output/sample_matrix_order.csv")
familyTax <- read.csv("output/sample_matrix_family.csv")
genusTax <- read.csv("output/sample_matrix_genus.csv")
```



```{r phylum, echo=FALSE}
water.phylum <- select(phylumTax, one_of(c("Phylum", water)))
rownames(water.phylum) <- water.phylum[,1]
water.phylum <- as.data.frame(t(water.phylum[,-1]))

copepod.phylum <- select(phylumTax, one_of(c("Phylum", copepod)))
rownames(copepod.phylum) <- copepod.phylum[,1]
copepod.phylum <- as.data.frame(t(copepod.phylum[,-1]))

water.phy.range <- range(water.phylum$Actinobacteria)
water.phy.median <- round(median(water.phylum$Actinobacteria), 2)

cop.phy.range <- range(copepod.phylum$Actinobacteria)
cop.phy.median <- round(median(copepod.phylum$Actinobacteria), 2)

cop.prot.range <- range(copepod.phylum$Proteobacteria)
cop.prot.median <- round(median(copepod.phylum$Proteobacteria), 2)

wat.prot.range <- range(water.phylum$Proteobacteria)
wat.prot.median <- round(median(water.phylum$Proteobacteria), 2)

wat.bact.range <- range(water.phylum$Bacteroidetes)
cop.bact.range <- range(copepod.phylum$Bacteroidetes)
```

Water and copepod bacterial communities are distinct even at the level of phyla. Water bacterial communities are dominated by Actinobacteria, which are significantly less common in copepod microbiomes (P =  ). Indeed, members of the Actinobacteria phylum comprise between `r water.phy.range[1]`% and `r water.phy.range[2]`% of the total water bacterial community (median = `r water.phy.median`%) while only comprising between `r cop.phy.range[1]`% and `r cop.phy.range[2]`% of the copepod community (median = `r cop.phy.median`%). Copepod communities, in turn, are dominated by Proteobacteria and to a lesser extent Bacteroidetes (FIG). Proteobacteria accounted for between `r cop.prot.range[1]`% and `r cop.prot.range[2]`% of copepods-associated samples (median = `r cop.prot.median`%) and between `r wat.prot.range[1]`% and `r wat.prot.range[2]`% of water samples (median = `r wat.prot.median`%). While Bacteroidetes abundance was relatively consistent across water samples (range `r wat.bact.range[1]`% - `r wat.bact.range[2]`%), Bacteroidetes abundance varied significantly in copepod microbiomes, with some locations containing as much as `r cop.bact.range[2]`% bacteroidetes while others as low as `r cop.bact.range[1]`%.

```{r class, echo=FALSE}
water.class <- select(classTax, one_of(c("Class", water)))
nam <- make.names(water.class[,1], unique=TRUE)
rownames(water.class) <- nam
water.class <- as.data.frame(t(water.class[,-1]))
wat.colmeans <- colMeans(water.class)
wat.colmeans <- order(wat.colmeans, decreasing=TRUE)

copepod.class <- select(classTax, one_of(c("Class", copepod)))
nam <- make.names(copepod.class[,1], unique=TRUE)
rownames(copepod.class) <- nam
copepod.class <- as.data.frame(t(copepod.class[,-1]))
cop.colmeans <- colMeans(copepod.class)

water.bet.range <- range(water.class$Betaproteobacteria)
water.bet.median <- round(median(water.class$Betaproteobacteria), 2)

cop.gam.range <- range(copepod.class$Gammaproteobacteria)
cop.gam.median <- round(median(copepod.class$Gammaproteobacteria), 2)


```

Of the proteobacteria Gammaproteobacteria was the most dominant class in copepod microbiomes, between `r cop.gam.range[1]`% and `r cop.gam.range[2]`% of the total taxa in copepod-associated samples (median = `r cop.gam.median`%). In water this was not the case, and the most dominant class of proteobacteria in water was Betaproteobacteria, between `r water.bet.range[1]`% and `r water.bet.range[2]`% of water samples. The pattern seems to be, however, that water-associated samples have more relative similarity at these higher taxonomic levels. While class abundance in water bacterial communities seemed to remain relatively constant, even between saltwater and freshwater samples, copepod samples did not show this pattern. The relative abundance of bacterial classes was much more variable between copepod-associated samples (FIG).

```{r level_comparison, echo=FALSE}
a <- as.list((apply(water.phylum, 1, function(x) length(which(x > 0)))))
b <- as.list((apply(copepod.phylum, 1, function(x) length(which(x > 0)))))
c <- data.frame(append(a, b))

#class
a <-as.list((apply(water.class, 1, function(x) length(which(x > 0)))))
b <- as.list((apply(copepod.class, 1, function(x) length(which(x > 0)))))
c <- rbind(c, data.frame(append(a, b)))

#order
water.order <- select(orderTax, one_of(c("Order", water)))
nam <- make.names(water.order[,1], unique=TRUE)
rownames(water.order) <- nam
water.order <- data.frame(t(water.order[,-1]))

copepod.order <- select(orderTax, one_of(c("Order", copepod)))
nam <- make.names(copepod.order[,1], unique=TRUE)
rownames(copepod.order) <- nam
copepod.order <- data.frame(t(copepod.order[,-1]))

a <-as.list((apply(water.order, 1, function(x) length(which(x > 0)))))
b <- as.list((apply(copepod.order, 1, function(x) length(which(x > 0)))))
c <- rbind(c, data.frame(append(a, b)))

#family
water.family <- select(familyTax, one_of(c("Family", water)))
nam <- make.names(water.family[,1], unique=TRUE)
rownames(water.family) <- nam
water.family <- data.frame(t(water.family[,-1]))

copepod.family <- select(familyTax, one_of(c("Family", copepod)))
nam <- make.names(copepod.family[,1], unique=TRUE)
rownames(copepod.family) <- nam
copepod.family <- data.frame(t(copepod.family[,-1]))

a <-as.list((apply(water.family, 1, function(x) length(which(x > 0)))))
b <- as.list((apply(copepod.family, 1, function(x) length(which(x > 0)))))
c <- rbind(c, data.frame(append(a, b)))

#genus
water.genus <- select(genusTax, one_of(c("Genus", water)))
nam <- make.names(water.genus[,1], unique=TRUE)
rownames(water.genus) <- nam
water.genus <- data.frame(t(water.genus[,-1]))

copepod.genus <- select(genusTax, one_of(c("Genus", copepod)))
nam <- make.names(copepod.genus[,1], unique=TRUE)
rownames(copepod.genus) <- nam
copepod.genus <- data.frame(t(copepod.genus[,-1]))

a <-as.list((apply(water.genus, 1, function(x) length(which(x > 0)))))
b <- as.list((apply(copepod.genus, 1, function(x) length(which(x > 0)))))
c <- rbind(c, data.frame(append(a, b)))

c <- data.frame(t(c))
tax <- c("Phylum", "Class","Order", "Family", "Genus")
colnames(c) <- tax
env <- c(rep("w", 14), rep("c", 14))
c <- cbind(c, Environment = env)
c <- cbind(c, Location = sapply(rownames(c), function(x) substr(x, 1, 2)))
```

## Taxonomic level abundance

```{r, echo=FALSE}
### Are there more Phylum in water samples than copepod? 
phy <- t.test(c$Phylum[1:12], c$Phylum[13:24], paired=TRUE, alt = "greater")

### Are there more Classes in water samples than copepod? 
cla <- t.test(c$Class[1:12], c$Class[13:24], paired=TRUE, alt = "greater")

### Are there more Orders in water samples than copepod? 
ord <- t.test(c$Order[1:12], c$Order[13:24], paired=TRUE, alt = "greater")

### Are there more Families in water samples than copepod? 
fam <- t.test(c$Family[1:12], c$Family[13:24], paired=TRUE, alt = "greater")

### Are there more Genera in water samples than copepod? 
gen <- t.test(c$Genus[1:12], c$Genus[13:24], paired=TRUE, alt = "greater")
```

### Manuscript Passage:

This same pattern holds for higher taxonomic levels; copepod microbiomes consist of fewer genera (P=`r as.character(signif(gen$p.value, 2))`), families (P=`r as.character(signif(fam$p.value, 2))`), orders (P=`r as.character(signif(ord$p.value, 2))`), classes (P=`r as.character(signif(cla$p.value, 2))`), and phyla (P=`r as.character(signif(phy$p.value, 2))`) than water samples from the same locations.
