==> manuscript_analyses/05_betaDiv_vst_figs.Rmd <==
---
output: html_document
---
#E affinis 16S analysis
### Martin Bontrager

```{r global_options, include=FALSE}
knitr::opts_chunk$set(fig.width=12, fig.height=8, fig.path='Figs_betaDiv_vstFigs/',
                      echo=FALSE, warning=FALSE, message=FALSE)
```

```{r read_data}
library(DESeq2)
library(phyloseq)
library(RColorBrewer)
library(ggplot2)
library(tidyr)
library(ape)
library(dplyr)
library(reshape2)
library(knitr)

set.seed(9373399)

otuTable <- read.csv("data/otus.csv", header = TRUE, row.names = 1)
taxTable <- read.csv("data/taxa.csv", quote = '"')
sampleData <- read.csv("data/sampleData.csv")
tree <- read.tree("data/tree.tre")
rownames(sampleData) <- sampleData[, 1]
sampleData <- sampleData[, -1]
taxmat <- as.matrix(taxTable)
rownames(taxmat) <- taxmat[, 1]
taxmat <- taxmat[, -1]


otuTable <- otu_table(otuTable, taxa_are_rows = FALSE)
taxTable <- tax_table(taxmat)

physeq <- phyloseq(otuTable, taxTable, tree)
sample_data(physeq) <- sampleData

## Variance Stabilized sample counts with DESeq2 (per "waste not, want not" paper)
pseq_deseq <- phyloseq_to_deseq2(physeq, ~Environment)

# calculate geometric means prior to estimate size factors
gm_mean = function(x, na.rm=TRUE){
  exp(sum(log(x[x > 0]), na.rm=na.rm) / length(x))
}
geoMeans <- apply(counts(pseq_deseq), 1, gm_mean)

#Dispersion Estimates in DESeq2
pseq_vst <- estimateSizeFactors(pseq_deseq, geoMeans = geoMeans)
pseq_vst <- estimateDispersions(pseq_vst)
pseq_vst <- getVarianceStabilizedData(pseq_vst)
pseq_vst[pseq_vst < 0] <- 0

otu_table(physeq) <- otu_table(pseq_vst, taxa_are_rows = TRUE)

rm(pseq_vst, geoMeans, pseq_deseq, sampleData, taxmat, otuTable, taxTable, tree)
```


```{r ggplottheme}
# ## ggplot theming
theme_set(theme_bw())
pal = "Set1"
scale_colour_discrete <- function(palname = pal, ...) {
        scale_colour_brewer(palette = palname, ...)
}

scale_fill_discrete <- function(palname = pal, ...) {
        scale_fill_brewer(palette = palname, ...)
}

getPalette <- colorRampPalette(brewer.pal(9, "Set1"))
```

# Beta diversity analyses

Since I have a phylogenetic tree for these data, I can use unifrac weighted and unweighted distance metrics which will be really great.

```{r calc_nonphylo_distances, fig.width=6, fig.height=4}
theme_set(theme_bw())
# Remove unifrac_weighted, unweighted and dpcoa distance metrics:
dist_methods <- unlist(distanceMethodList)
dist_methods <- dist_methods[-c(1:3, 5:7, 9, 11:13, 16:17, 19:47)]

plist <- vector("list", length(dist_methods))
names(plist) = dist_methods

for( i in dist_methods ){
    iDist <- distance(physeq, method=i)
    iMDS  <- ordinate(physeq, "MDS", distance=iDist)
    ## Make plot
    # Don't carry over previous plot (if error, p will be blank)
    p <- NULL
    # Create plot, store as temp variable, p
    p <- plot_ordination(physeq, iMDS, color="Environment")
    # Add title to each plot
    p <- p + ggtitle(paste("MDS using distance method ", i, sep="")) +
            geom_point(size = 6, alpha = 0.9)
    # Remove Initial points from plot
    p$layers <- p$layers[-1]
    # Save the graphic to file.
    plist[[i]] = p
}

require(plyr)
df <- ldply(plist, function(x) x$data)
names(df)[1] <- "distance"
p <- ggplot(df, aes(Axis.1, Axis.2, color=Environment)) + 
            geom_point(size=2.5, alpha=0.9) +
            facet_wrap(~distance, scales="free") +
            theme(strip.text.x = element_text(size = 14),
                  axis.title.x = element_blank(),
                  axis.title.y = element_blank(),
                  axis.text.x = element_blank(),
                  axis.text.y = element_blank())
p
pdf("Fig_beta_div_environment.pdf", height=4, width=6)
p
dev.off()
```

It's best to look at a subset of these instead of all the distance matrices. Here are the PCoA ordination plots for Jaccard (Presence/absence only), Bray-Curtis (abundance-based), and Jenson-Shannon distances:

```{r distance_methods, fig.width=6, fig.height=4}
theme_set(theme_bw())
colorCount = length(levels(sample_data(physeq)$Location))

c.jaccard <- ordinate(physeq, "PCoA", "jaccard")
c.bray <- ordinate(physeq, "PCoA", "bray")
c.jsd <- ordinate(physeq, "PCoA", "jsd")
plot_ordination(physeq, c.jaccard, color = "Location", shape = "Environment") +
    ggtitle(paste("PCoA w/ jaccard Distance")) +
    geom_point(size = 6, alpha = 0.9) +
    scale_color_manual(values = getPalette(colorCount))

plot_ordination(physeq, c.jaccard, axes = c(3, 4), color = "Location", 
                shape = "Environment") +
    ggtitle(paste("PCoA w/ jaccard Distance on axis 3, 4")) +
    geom_point(size = 6, alpha = 0.9) +
    scale_color_manual(values = getPalette(colorCount))

plot_ordination(physeq, c.bray, color = "Location", shape = "Environment") +
    ggtitle(paste("PCoA w/ Bray-Curtis Distance")) +
    geom_point(size = 6, alpha = 0.9) +
    scale_color_manual(values = getPalette(colorCount))

plot_ordination(physeq, c.bray, axes = c(3, 4), color = "Location", 
                shape = "Environment") +
    ggtitle(paste("PCoA w/ Bray-Curtis Distance on axis 3, 4")) +
    geom_point(size = 6, alpha = 0.9) +
    scale_color_manual(values = getPalette(colorCount))

plot_ordination(physeq, c.jsd, color = "Location", shape = "Environment") +
    ggtitle(paste("PCoA w/ jsd Distance")) +
    geom_point(size = 6, alpha = 0.9) +
    scale_color_manual(values = getPalette(colorCount))

plot_ordination(physeq, c.jsd, axes = c(3, 4), color = "Location", 
                shape = "Environment") +
    ggtitle(paste("PCoA w/ jsd Distance on axis 3, 4")) +
    geom_point(size = 6, alpha = 0.9) +
    scale_color_manual(values = getPalette(colorCount))
```

# Here are those same distances plotted with NMDS ordination

```{r NMDS, fig.width=6, fig.height=4}
c.jaccard <- ordinate(physeq, "NMDS", "jaccard")
c.bray <- ordinate(physeq, "NMDS", "bray")
c.jsd <- ordinate(physeq, "NMDS", "jsd")
plot_ordination(physeq, c.jaccard, color = "Location", shape = "Environment") +
    ggtitle(paste("NMDS w/ jaccard Distance")) +
    geom_point(size = 6, alpha = 0.9) +
    scale_color_manual(values = getPalette(colorCount))

plot_ordination(physeq, c.bray, color = "Location", shape = "Environment") +
    ggtitle(paste("NMDS w/ Bray-Curtis Distance")) +
    geom_point(size = 6, alpha = 0.9) +
    scale_color_manual(values = getPalette(colorCount))

plot_ordination(physeq, c.bray, color = "WaterType", shape = "Environment") +
    ggtitle(paste("NMDS w/ Bray-Curtis Distance")) +
    geom_point(size = 6, alpha = 0.9)

plot_ordination(physeq, c.bray, color="HostxWater") +
    ggtitle(paste("NMDS w/ Bray-Curtis Distance")) +
    geom_point(size = 6, alpha = 0.9) +
    scale_color_manual(values=c("#E41A1C", "#984EA3", "#377EB8", "#4DAF4A")) +
    theme(axis.title.y = element_text(size=20),
          axis.title.x = element_text(size=20),
          plot.title = element_text(size=24, face="bold", vjust=2))

plot_ordination(physeq, c.jsd, color = "Location", shape = "Environment") +
    ggtitle(paste("NMDS w/ jsd Distance")) +
    geom_point(size = 6, alpha = 0.9) +
    scale_color_manual(values = getPalette(colorCount))

plot_ordination(physeq, c.jsd, color = "WaterType", shape = "Environment") +
    ggtitle(paste("NMDS w/ jsd Distance")) +
    geom_point(size = 6, alpha = 0.9)

```

# UniFrac Distances

This method relies on the calculation of distances via shared phylogenetic branch-length between samples on a phylogenetic tree. Shown below are weighted UniFrac distances (abundance is considered)


```{r unifrac, fig.width=6, fig.height=4}
theme_set(theme_bw())
# Find color count (sample location)
colorCount = length(levels(sample_data(physeq)$Location))
#Unifrac weighted distance
unifrac_weighted <- UniFrac(physeq, weighted = TRUE)
# Ordinate
iMDS  <- ordinate(physeq, "NMDS", distance=unifrac_weighted)
plot_ordination(physeq, iMDS, color="Location", shape = "Environment") +
    ggtitle(paste("NMDS w/ UNIFRAC weighted")) +
    geom_point(size = 6, alpha = 0.9) +
    scale_color_manual(values = getPalette(colorCount))

#Plot Unifrac with hostxwater colors,
plot_ordination(physeq, iMDS, color="HostxWater") +
    ggtitle(paste("NMDS w/ UNIFRAC weighted")) +
    geom_point(size = 6, alpha = 0.9) +
    scale_color_manual(values=c("#E41A1C", "#984EA3", "#377EB8", "#4DAF4A")) +
    theme(axis.title.y = element_text(size=20),
          axis.title.x = element_text(size=20),
          plot.title = element_text(size=24, face="bold", vjust=2))

iMDS  <- ordinate(physeq, "PCoA", distance=unifrac_weighted)
colorCount = length(levels(sample_data(physeq)$Location))

#Plot Unifrac with Cladexwater colors,
plot_ordination(physeq, iMDS, color="Location", shape="Environment") +
    ggtitle(paste("PCoA w/ UNIFRAC weighted")) +
    geom_point(size = 6, alpha = 0.9) +
    scale_color_manual(values = getPalette(colorCount)) +
    theme(axis.title.y = element_text(size=20),
          axis.title.x = element_text(size=20),
          plot.title = element_text(size=24, face="bold", vjust=2))
```

#UniFrac Unweighted

Generally speaking I prefer to use unweighted Unifrac, which calculates distance based solely on presence/absence rather than overall abundance. They are both probably good measures, I just feel like 16S amplification bias and other sample prep/extraction issues might confuse the weighted.

```{r ordination_plot_unifrac_un, fig.width=6, fig.height=4}
theme_set(theme_bw())
colorCount = length(levels(sample_data(physeq)$Location))

#Unifrac unweighted distance
unifrac_unweighted <- UniFrac(physeq, weighted = FALSE)

#Plot Unifrac with Environment colors,
iMDS  <- ordinate(physeq, "MDS", distance=unifrac_unweighted)
pdf("FigS6-UniFrac_MDS_unweighted.pdf", height=4, width=6)
plot_ordination(physeq, iMDS, color="Environment") +
    ggtitle(paste("MDS w/ UNIFRAC unweighted")) +
    geom_point(size = 2.5, alpha = 1) +
    theme(axis.title.y = element_text(size=20),
          axis.title.x = element_text(size=20),
          plot.title = element_blank())
dev.off()


# Ordinate
iMDS  <- ordinate(physeq, "NMDS", distance=unifrac_unweighted)

#Unifrac unweighted plot
plot_ordination(physeq, iMDS, color="Location") +
    ggtitle(paste("NMDS w/ UNIFRAC unweighted")) +
    geom_point(size = 6, alpha = 0.9) +
    theme(axis.title.y = element_text(size=20),
          axis.title.x = element_text(size=20),
          plot.title = element_text(size=24, face="bold", vjust=2)) +
    scale_color_manual(values = getPalette(colorCount))

theme_set(theme_bw(base_size = 22))
#Plot Unifrac with hostxwater colors,
pdf("UniFrac_unweighted_NMDS.pdf", width=6, height=4)
plot_ordination(physeq, iMDS, color="HostxWater", shape="Environment") +
    geom_point(size = 2.5) +
    scale_color_manual(values=c("lightpink2", "darkred", "lightblue", "darkblue")) +
    guides(shape=FALSE, color=FALSE)
    theme(axis.title.y = element_text(size=20),
          axis.title.x = element_text(size=20),
          plot.title = element_blank())
dev.off()

# PCoA plots

iMDS  <- ordinate(physeq, "PCoA", distance=unifrac_unweighted)

#Plot Unifrac with hostxwater colors,
plot_ordination(physeq, iMDS, color="HostxWater") +
    ggtitle(paste("PCoA w/ UNIFRAC unweighted")) +
    geom_point(size = 6, alpha = 0.9) +
    scale_color_manual(values=c("#E41A1C", "#984EA3", "#377EB8", "#4DAF4A")) +
    theme(axis.title.y = element_text(size=20),
          axis.title.x = element_text(size=20),
          plot.title = element_text(size=24, face="bold", vjust=2))
```

# Copepods only

```{r unifrac_ngnw}
theme_set(theme_bw())
# Find color count (sample location)
ngnw <- subset_samples(physeq, Environment != "Water")
colorCount = length(levels(sample_data(ngnw)$Location))

#UniFrac unweighted distances
unifrac_unweighted <- UniFrac(ngnw, weighted = FALSE)

# Ordinate
iMDS  <- ordinate(ngnw, "NMDS", distance=unifrac_unweighted)

#Unifrac unweighted plot
plot_ordination(ngnw, iMDS, color="Location") +
    ggtitle(paste("NMDS w/ UNIFRAC unweighted")) +
    geom_point(size = 6, alpha = 0.9) +
    theme(axis.title.y = element_text(size=20),
          axis.title.x = element_text(size=20),
          plot.title = element_text(size=24, face="bold", vjust=2)) +
    scale_color_manual(values = getPalette(colorCount))

#Plot Unifrac with hostxwater colors,
plot_ordination(ngnw, iMDS, color="HostxWater") +
    ggtitle(paste("NMDS w/ UNIFRAC unweighted")) +
    geom_point(size = 6, alpha = 0.9) +
    scale_color_manual(values=c("#E41A1C", "#984EA3", "#377EB8", "#4DAF4A")) +
    theme(axis.title.y = element_text(size=20),
          axis.title.x = element_text(size=20),
          plot.title = element_blank())

theme_set(theme_bw(base_size = 22))
#Plot Unifrac with hostxwater colors,
pdf("fig-NMDS_copepodsunifrac_unweighted.pdf", width=7, height=5)
#png("UniFrac_unweighted_NMDS.png", width=800, height=600)
plot_ordination(ngnw, iMDS, color="WaterType") +
    geom_point(size = 5) +
    scale_color_manual(values=c("lightpink2", "darkred")) +
    guides(shape=FALSE, color=FALSE)
    theme(axis.title.y = element_text(size=20),
          axis.title.x = element_text(size=20),
          plot.title = element_blank())
dev.off()
```
